+++ CHANGELOG.md
@@ -7,7 +7,8 @@ x.y.z Release notes (yyyy-MM-dd)
 
 ### Fixed
 * <How to hit and notice issue? what was the impact?> ([#????](https://github.com/realm/realm-cocoa/issues/????), since v?.?.?)
-* None.
+* Fix the spelling of `ObjectKeyIdentifiable`. The old spelling is available
+  and deprecated for compatiblity.
 
 <!-- ### Breaking Changes - ONLY INCLUDE FOR NEW MAJOR version -->